+++ module/CLI/src/Command/ShortUrl/ListShortUrlsCommand.php
@@ -78,6 +78,18 @@ class ListShortUrlsCommand extends Command
                 'If tags is provided, returns only short URLs having ALL tags.',
             )
             ->addOption(
+                'exclude-max-visits-reached',
+                null,
+                InputOption::VALUE_NONE,
+                'Excludes short URLs which reached their max amount of visits.',
+            )
+            ->addOption(
+                'exclude-past-valid-until',
+                null,
+                InputOption::VALUE_NONE,
+                'Excludes short URLs which have a "validUntil" date in the past.',
+            )
+            ->addOption(
                 'order-by',
                 'o',
                 InputOption::VALUE_REQUIRED,
@@ -133,6 +145,8 @@ class ListShortUrlsCommand extends Command
             ShortUrlsParamsInputFilter::ORDER_BY => $orderBy,
             ShortUrlsParamsInputFilter::START_DATE => $startDate?->toAtomString(),
             ShortUrlsParamsInputFilter::END_DATE => $endDate?->toAtomString(),
+            ShortUrlsParamsInputFilter::EXCLUDE_MAX_VISITS_REACHED => $input->getOption('exclude-max-visits-reached'),
+            ShortUrlsParamsInputFilter::EXCLUDE_PAST_VALID_UNTIL => $input->getOption('exclude-past-valid-until'),
         ];
 
         if ($all) {