+++ module/Core/test/ShortUrl/Entity/ShortUrlTest.php
@@ -38,7 +38,7 @@ class ShortUrlTest extends TestCase
     public function provideInvalidShortUrls(): iterable
     {
         yield 'with custom slug' => [
-            ShortUrl::fromMeta(ShortUrlCreation::fromRawData(['customSlug' => 'custom-slug', 'longUrl' => ''])),
+            ShortUrl::create(ShortUrlCreation::fromRawData(['customSlug' => 'custom-slug', 'longUrl' => ''])),
             'The short code cannot be regenerated on ShortUrls where a custom slug was provided.',
         ];
         yield 'already persisted' => [
@@ -77,7 +77,7 @@ class ShortUrlTest extends TestCase
      */
     public function shortCodesHaveExpectedLength(?int $length, int $expectedLength): void
     {
-        $shortUrl = ShortUrl::fromMeta(ShortUrlCreation::fromRawData(
+        $shortUrl = ShortUrl::create(ShortUrlCreation::fromRawData(
             [ShortUrlInputFilter::SHORT_CODE_LENGTH => $length, 'longUrl' => ''],
         ));